requirements.txt
amqp==2.4.2
billiard==3.6.0.0
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
kombu==4.5.0
pika==1.0.1
psycopg2==2.8.2
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
vine==1.3.0
